requirements.txt
boto3
requests
pyyaml
tabulate
jinja2
packaging
pytest
pytest-cov
